#10 move knowledge graph controller to new file kgcontroller.py
